+++ mediadevice/device.h
@@ -45,7 +45,7 @@ public:
 
     virtual bool isReadyToOpen() const = 0;
     bool waitReadyToOpen(const unsigned long &timeout=30000);
-    bool waitOpen(const int &timeout=30000);
+    bool waitOpen(const unsigned long &timeout=30000);
 
     virtual QByteArray read(qint64 maxlen) = 0;
